fix(test): cleanup adapter tests
